+++ projects/web/src/components/Sidebar/SidebarContainer.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import {connect} from 'react-redux'
+import Sidebar from './Sidebar.jsx'
+import {toggleSidebar} from './actions'
+import {withRouter} from 'react-router-dom'
+
+const mapStateToProps = state => ({
+    isSidebarOpen: state.Sidebar.isSidebarOpen
+});
+
+const mapDispatchToProps = dispatch => ({
+    toggleSidebar: () => {
+        dispatch(toggleSidebar())
+    }
+});
+
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(Sidebar));