+++ projects/web/tools/webpack/webpack.loaders.config.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const reStyle = /\.(css|less|scss|sss)$/;
+const publicPath = require('./publicPath');
 
 const getLoaders = (bundle, mode) => {
 
@@ -88,6 +89,7 @@ const getLoaders = (bundle, mode) => {
                 limit: 10000,
                 mimetype: 'application/font-woff',
                 name: 'fonts/[name].[ext]',
+                publicPath: !_.isEmpty(publicPath) ? publicPath : '/',
                 emitFile: bundle === 'client'
             }
         },
@@ -96,6 +98,7 @@ const getLoaders = (bundle, mode) => {
             loader: 'file-loader',
             options: {
                 name: 'fonts/[name].[ext]',
+                publicPath: !_.isEmpty(publicPath) ? publicPath : '/',
                 emitFile: bundle === 'client'
             }
         },
@@ -104,6 +107,7 @@ const getLoaders = (bundle, mode) => {
             loader: 'file-loader',
             options: {
                 name: 'img/[name].[ext]',
+                publicPath: !_.isEmpty(publicPath) ? publicPath : '/',
                 emitFile: bundle === 'client'
             }
         }